Refactoring telemetry integration in handler (#55)
* Refactoring telemetry integration in handler

* Adding Telemetry on/off switch for handler

* Removing unused telemetry code from Logger

* Setting timestamp in OperationId within telemetry events for all non-enable commands

* Addressing PR comments

* Handler: Logger to add events to telemetry

* minor changes

+++ src/extension/src/TelemetryWriter.py
@@ -27,42 +27,46 @@ from extension.src.Constants import Constants
 class TelemetryWriter(object):
     """Class for writing telemetry data to events"""
 
-    def __init__(self):
+    def __init__(self, logger):
+        self.logger = logger
         self.events_folder_path = None
-        self.operation_id = ""
+        self.__operation_id = ""
 
-    def __new_event_json(self, task_name, event_level, message):
+    def __new_event_json(self, event_level, message, task_name):
         return {
             "Version": Constants.EXT_VERSION,
-            "Timestamp": str((datetime.datetime.utcnow()).strftime(Constants.UTC_DATETIME_FORMAT)),
+            "Timestamp": str(datetime.datetime.utcnow()),
             "TaskName": task_name,
             "EventLevel": event_level,
             "Message": self.__ensure_message_restriction_compliance(message),
             "EventPid": "",
             "EventTid": "",
-            "OperationId": self.operation_id  # we can provide activity id from config settings here, but currently we only read settings file for enable command
+            "OperationId": self.__operation_id  # This should have activity id from from config settings, but since we only read settings file for enable command, enable command will have activity id set here and all non-enable commands will have this as a timestamp
         }
 
-    @staticmethod
-    def __ensure_message_restriction_compliance(full_message):
+    def __ensure_message_restriction_compliance(self, full_message):
         """ Removes line breaks, tabs and restricts message to a byte limit """
         message_size_limit_in_bytes = Constants.TELEMETRY_MSG_SIZE_LIMIT_IN_BYTES
         formatted_message = re.sub(r"\s+", " ", str(full_message))
 
         if len(formatted_message.encode('utf-8')) > message_size_limit_in_bytes:
+            self.logger.log_telemetry_module("Data sent to telemetry will be truncated as it exceeds size limit. [Message={0}]".format(str(formatted_message)))
             formatted_message = formatted_message.encode('utf-8')
             bytes_dropped = len(formatted_message) - message_size_limit_in_bytes + Constants.TELEMETRY_BUFFER_FOR_DROPPED_COUNT_MSG_IN_BYTES
             return formatted_message[:message_size_limit_in_bytes - Constants.TELEMETRY_BUFFER_FOR_DROPPED_COUNT_MSG_IN_BYTES].decode('utf-8') + '. [{0} bytes dropped]'.format(bytes_dropped)
 
         return formatted_message
 
-    def write_event(self, task_name, message, event_level=Constants.TelemetryEventLevel.Informational):
+    def write_event(self, message, event_level=Constants.TelemetryEventLevel.Informational, task_name=Constants.TELEMETRY_TASK_NAME):
         """ Creates and writes event to event file after validating none of the telemetry size restrictions are breached """
+        if self.events_folder_path is None or not os.path.exists(self.events_folder_path) or not Constants.TELEMETRY_ENABLED_AT_EXTENSION:
+            return
+
         self.__delete_older_events()
 
-        event = self.__new_event_json(task_name, event_level, message)
+        event = self.__new_event_json(event_level, message, task_name)
         if len(json.dumps(event)) > Constants.TELEMETRY_EVENT_SIZE_LIMIT_IN_BYTES:
-            print("Cannot send data to telemetry as it exceeded the acceptable data size. [Data not sent={0}]".format(json.dumps(message)))
+            self.logger.log_telemetry_module_error("Cannot send data to telemetry as it exceeded the acceptable data size. [Data not sent={0}]".format(json.dumps(message)))
         else:
             self.write_event_using_temp_file(self.events_folder_path, event)
 
@@ -72,7 +76,7 @@ class TelemetryWriter(object):
             # Not deleting any existing event files as the event directory does not exceed max limit. At least one new event file can be added. Not printing this statement as it will add repetitive logs
             return
 
-        print("Events directory size exceeds maximum limit. Deleting older event files until at least one new event file can be added.")
+        self.logger.log_telemetry_module("Events directory size exceeds maximum limit. Deleting older event files until at least one new event file can be added.")
         event_files = [os.path.join(self.events_folder_path, event_file) for event_file in os.listdir(self.events_folder_path) if (event_file.lower().endswith(".json"))]
         event_files.sort(key=os.path.getmtime, reverse=True)
 
@@ -84,9 +88,9 @@ class TelemetryWriter(object):
 
                 if os.path.exists(event_file):
                     os.remove(event_file)
-                    print("Deleted event file. [File={0}]".format(repr(event_file)))
+                    self.logger.log_telemetry_module("Deleted event file. [File={0}]".format(repr(event_file)))
             except Exception as e:
-                print("Error deleting event file. [File={0}] [Exception={1}]".format(repr(event_file), repr(e)))
+                self.logger.log_telemetry_module_error("Error deleting event file. [File={0}] [Exception={1}]".format(repr(event_file), repr(e)))
 
         if self.__get_events_dir_size() >= Constants.TELEMETRY_DIR_SIZE_LIMIT_IN_BYTES:
             raise Exception("Older event files were not deleted. Current event will not be sent to telemetry as events directory size exceeds maximum limit")
@@ -111,10 +115,10 @@ class TelemetryWriter(object):
                 tempname = tf.name
             shutil.move(tempname, file_path)
         except Exception as error:
-            raise Exception("Unable to write to {0}. Error: {1}.".format(str(file_path), repr(error)))
+            raise Exception("Unable to write to telemetry. [Event File={0}] [Error={1}].".format(str(file_path), repr(error)))
 
     def set_operation_id(self, operation_id):
-        self.operation_id = operation_id
+        self.__operation_id = operation_id
 
     def __get_events_dir_size(self):
         return sum([os.path.getsize(os.path.join(self.events_folder_path, f)) for f in os.listdir(self.events_folder_path) if os.path.isfile(os.path.join(self.events_folder_path, f))])
